Move 'artifactory_artifact' resource to own directory
