+++ pkg/artifactory/provider/framework.go
@@ -16,7 +16,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	datasource_artifact "github.com/jfrog/terraform-provider-artifactory/v11/pkg/artifactory/datasource/artifact"
 	datasource_repository "github.com/jfrog/terraform-provider-artifactory/v11/pkg/artifactory/datasource/repository"
-	rs "github.com/jfrog/terraform-provider-artifactory/v11/pkg/artifactory/resource"
+	"github.com/jfrog/terraform-provider-artifactory/v11/pkg/artifactory/resource/artifact"
 	"github.com/jfrog/terraform-provider-artifactory/v11/pkg/artifactory/resource/configuration"
 	"github.com/jfrog/terraform-provider-artifactory/v11/pkg/artifactory/resource/lifecycle"
 	"github.com/jfrog/terraform-provider-artifactory/v11/pkg/artifactory/resource/security"
@@ -197,7 +197,7 @@ func (p *ArtifactoryProvider) Configure(ctx context.Context, req provider.Config
 // Resources satisfies the provider.Provider interface for ArtifactoryProvider.
 func (p *ArtifactoryProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
-		rs.NewArtifactResource,
+		artifact.NewArtifactResource,
 		user.NewAnonymousUserResource,
 		user.NewManagedUserResource,
 		user.NewUnmanagedUserResource,
