+++ pkg/artifactory/resource/artifact/resource_artifactory_artifact.go
@@ -1,4 +1,4 @@
-package resource
+package artifact
 
 import (
 	"context"
