+++ pkg/artifactory/resource/artifact/resource_artifactory_artifact_test.go
@@ -1,4 +1,4 @@
-package resource_test
+package artifact_test
 
 import (
 	"fmt"
